Fixed pawn, knight, king movement functions. Fixed unit tests. Added react client scaffold.

+++ system/mongoose.js
@@ -2,24 +2,26 @@
 var mongoose = require('mongoose');
 
 mongoose.Promise = global.Promise;
-mongoose.connect('mongodb://localhost:27017/VinylMateDB', {
+mongoose.connect('mongodb://localhost:27017/BrowserChessDB', {
   useNewUrlParser: true
 });
 
-before((done) => {
-  console.log('Connecting to test DB');
-  mongoose
-    .connect('mongodb://localhost:27017/VinylMateDBTest', {
-      useNewUrlParser: true
-    })
-    .then(() => {
-      console.log('Connected to test DB.');
-      done();
-    });
-});
+if (typeof before !== typeof undefined) {
+  before((done) => {
+    console.log('Connecting to test DB');
+    mongoose
+      .connect('mongodb://localhost:27017/BrowserChessDBTest', {
+        useNewUrlParser: true
+      })
+      .then(() => {
+        console.log('Connected to test DB.');
+        done();
+      });
+  });
 
-beforeEach((done) => {
-  mongoose.connection.db.dropDatabase().then(() => done());
-});
+  beforeEach((done) => {
+    mongoose.connection.db.dropDatabase().then(() => done());
+  });
+}
 
 module.exports = { mongoose };
